# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShadowNet API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /relay/nodes/{nodeId}/connect:
    post:
      operationId: connectToRelayNode
      tags: [relay]
      summary: Connect to relay node
      description: Initiates connection to a specific relay node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connection established
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelayConnection"
        "404":
          description: Node not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    RelayConnection:
      type: object
      properties:
        sessionId:
          type: string
        nodeId:
          type: string
        connectedAt:
          type: string
        maskedIp:
          type: string
        status:
          type: string
          enum: [connected, failed, pending]
      required:
        - sessionId
        - nodeId
        - connectedAt
        - maskedIp
        - status
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        retryAfter:
          type: integer
          description: Seconds to wait before retrying (only set on 429 responses)
      required:
        - error

    # ── Intelligence: X scanner schemas ───────────────────────────────────────
...